more work on query result validation
